Commit missed files
